requirements.txt
# 一年365赢项目依赖

# 核心依赖
requests>=2.28.0
PyYAML>=6.0
schedule>=1.2.0

# 网络爬虫相关
feedparser>=6.0.10
beautifulsoup4>=4.12.0
lxml>=4.9.0

# 数据处理
pandas>=2.0.0
numpy>=1.24.0

# 开发工具
pytest>=7.0.0
pytest-cov>=4.0.0
flake8>=6.0.0
black>=23.0.0

# 安全工具
gitleaks>=8.0.0

# 可选依赖（用于高级功能）
# openai>=0.27.0  # 如果需要其他AI API
# tweepy>=4.0.0  # 如果需要Twitter/X集成